// app/directives/helpers.d.ts
export declare function saveCache(url: any, data: any): void;
export declare function loadCache(url: any): any;
export declare function getParameterByName(name: any): string;
export declare function cleanQuerystring(param?: any, value?: any): void;
export declare function GooglelogOut(mess: any): void;
export declare function parseXml(xmlStr: any): any;
export declare function addp(url: string, param: string, value?: any): string;
export declare function fullapplink(site: any, ticketkey: any, inst: any, org: any): void;
export declare function mailtolink(ticketkey: any, inst: any, org: any): void;
export declare function htmlEscape(str: any): string;
export declare function linebreaks(value: any, args: any): any;
export declare const FileUrlHelper: {
    isPhonegap: () => boolean;
    ReplaceAll: (note: any, find: any, replace: any) => any;
    checkURL: (url: any) => boolean;
    matchKey: (search: any, array: any) => string;
    addUrls: (note: any, files: any) => any;
    getFileLink: (file: any, name: any) => string;
};
export declare function openURL(urlString: any): Window;
export declare function openURLsystem(urlString: any): any;
export declare function symbolEscape(str: any): string;
export declare function getPickerDateTimeFormat(showmonth?: any, istime?: any): string;
export declare function getDateTime(date: any, showmonth?: any, istime?: any): any;
export declare function getCurrency(value: any): string;
export declare function getFullName(firstname: any, lastname: any, email: any, name?: any): string;
export declare function toTitleCase(str: any): any;
export declare function getappTrackConversion(id: any): void;
export declare function spicePixelTrackConversion(): void;